% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_between_groups.R
\name{boxplot_between_groups}
\alias{boxplot_between_groups}
\title{Creating a boxplot and csv table based on the table with the get_RSCU function and grouping_table}
\usage{
boxplot_between_groups(
  get_RSCU_out,
  grouping_table,
  width,
  height,
  xlab,
  res,
  p.adjust.method = "bonferroni"
)
}
\arguments{
\item{get_RSCU_out}{the table was created using the get_RSCU function}

\item{grouping_table}{the table with Species and group column names}

\item{width}{width of charts}

\item{height}{height of charts}

\item{xlab}{x axis title}

\item{res}{resolution of charts}

\item{p.adjust.method}{the selection of a p-value correction method: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr"}
}
\value{
A plot and csv objects.
}
\description{
Creating a boxplot and csv table based on the table with the get_RSCU function and grouping_table
}
