% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RSCU_other.R
\name{get_codon_table}
\alias{get_codon_table}
\title{Retrieving the codon table}
\usage{
get_codon_table(codon_table_id)
}
\arguments{
\item{codon_table_id}{An integer specifying the codon table.}
}
\value{
A list with the element genetic_code - a vector of amino acids corresponding to codons.
}
\description{
This function returns the codon table for a specified genetic code.
}
