% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matrix.R
\name{get_matrix}
\alias{get_matrix}
\title{Converting the table from the get_RSCU function to a matrix}
\usage{
get_matrix(get_RSCU_out = get_RSCU_out)
}
\arguments{
\item{get_RSCU_out}{the table was created using the get_RSCU function}
}
\value{
A `matrix` object containing matrix_RSCU.
}
\description{
Converting the table from the get_RSCU function to a matrix
}
