% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RSCU_other.R
\name{get_RSCU_other}
\alias{get_RSCU_other}
\title{Calculating Relative Synonymous Codon Usage (RSCU)}
\usage{
get_RSCU_other(merged_sequences = "", codon_table_id = 1, pseudo_count = 1)
}
\arguments{
\item{merged_sequences}{A path to sequences or a list of sequences.}

\item{codon_table_id}{An integer specifying the codon table (1-6,9-16,21-26).}

\item{pseudo_count}{A value added to codon counts to avoid division by zero.}
}
\value{
A data frame with columns: codon, amino acid, eff, RSCU.
}
\description{
This function calculates RSCU values for codons in a given nucleotide sequence.
}
