% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_RSCU.R
\name{histogram_RSCU}
\alias{histogram_RSCU}
\title{Creating a barplot based on the table with the get_RSCU function}
\usage{
histogram_RSCU(get_RSCU_out = get_RSCU_out, title = "")
}
\arguments{
\item{get_RSCU_out}{the table was created using the get_RSCU function}

\item{title}{chart title}
}
\value{
A plot object.
}
\description{
Creating a barplot based on the table with the get_RSCU function
}
