% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitogenome_sequence.R
\docType{data}
\name{mitogenome_sequence}
\alias{mitogenome_sequence}
\title{FASTA sequences corresponding to samples_table}
\format{
A set of FASTA files.
}
\source{
Internal dataset for microbiome analysis.
}
\usage{
data(mitogenome_sequence)
}
\value{
A fasta list.
}
\description{
This dataset consists of FASTA files containing mitogenome gene sequences corresponding to the samples listed in the samples_table.
Each FASTA file is named according to the ID in the samples_table.
}
\examples{
data(mitogenome_sequence)
print(mitogenome_sequence)
}
\keyword{datasets}
