% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutrality_pr2.R
\name{neutrality_pr2}
\alias{neutrality_pr2}
\title{Creating a PR2 plot and neutrality plot based on the table with the get_RSCU function}
\usage{
neutrality_pr2(
  get_RSCU_out = get_RSCU_out,
  select = "",
  grouping_table = grouping_table
)
}
\arguments{
\item{get_RSCU_out}{the table was created using the get_RSCU function}

\item{select}{the "PR2_plot" or "neutrality_plot" output}

\item{grouping_table}{the table with Species and group column names}
}
\value{
A plot object and table object.
}
\description{
Creating a PR2 plot and neutrality plot based on the table with the get_RSCU function
}
