% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_fasta.R
\name{prepare_fasta}
\alias{prepare_fasta}
\title{Preparing fasta files for further processing}
\usage{
prepare_fasta(samples_table = samples_table, path = "", file_out = "")
}
\arguments{
\item{samples_table}{table with sequence descriptions}

\item{path}{path to the original sequence file}

\item{file_out}{path to the file with processed sequences}
}
\value{
None (invisible)
}
\description{
Preparing fasta files for further processing
}
