% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RSCU_other.R
\name{seq_to_data_frame}
\alias{seq_to_data_frame}
\title{Convert FASTA sequence to a data frame}
\usage{
seq_to_data_frame(merged_sequences = "")
}
\arguments{
\item{merged_sequences}{path to the file with prepared sequences via the prepare_fasta() function}
}
\value{
A data frame with columns: names, sequence
}
\description{
This function converts a FASTA sequence into a data frame.
}
