% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RSCU_other.R
\name{calculate_rscu}
\alias{calculate_rscu}
\title{Calculating Relative Synonymous Codon Usage (RSCU)}
\usage{
calculate_rscu(nucleotide_input, codon_table_id = 1, pseudo_count = 1)
}
\arguments{
\item{codon_table_id}{An integer specifying the codon table (1-6,9-16,21-26).}

\item{pseudo_count}{A value added to codon counts to avoid division by zero.}

\item{nucleotide_string}{A character string representing the nucleotide sequence.}
}
\value{
A data frame with columns: codon, amino acid, eff, RSCU.
}
\description{
This function calculates RSCU values for codons in a given nucleotide sequence.
}
