% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{Creating a csv table, boxplots and scatter plots table based on the table with the get_RSCU function}
\usage{
correlation(get_RSCU_out, Species_x, Species_y, xlab, ylab)
}
\arguments{
\item{get_RSCU_out}{the table was created using the get_RSCU function}

\item{Species_x}{The name from the get_RSCU table that you want to use to calculate the correlation will be on the x-axis}

\item{Species_y}{The name from the get_RSCU table that you want to use to calculate the correlation will be on the y-axis}

\item{xlab}{x axis title}

\item{ylab}{y axis titles}
}
\value{
A list with statistics and plot.
}
\description{
Creating a csv table, boxplots and scatter plots table based on the table with the get_RSCU function
}
