% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matrix.R
\name{get_matrix}
\alias{get_matrix}
\title{Convert RSCU Results to Analysis-Ready Matrix}
\usage{
get_matrix(get_RSCU_out = get_RSCU_out)
}
\arguments{
\item{get_RSCU_out}{A data frame generated by \code{\link{get_RSCU}}, 
\code{\link{get_RSCU_other}} or \code{\link{get_RSCU_other2}} containing
RSCU values for multiple samples.}
}
\value{
A list containing two elements:
\describe{
  \item{matrix}{Numeric matrix with samples as rows (rownames = sample IDs) 
    and codon-AA pairs as columns (format: "AA | codon")}
  \item{df_matrix}{Data frame version of the matrix with an additional 
    "Sample" column}
}
}
\description{
Transforms RSCU (Relative Synonymous Codon Usage) results from long to wide format,
creating a matrix suitable for multivariate analysis and visualization.
}
\details{
This function restructures RSCU data into a format compatible with:
\itemize{
  \item Multivariate statistics (PCA, clustering)
  \item Machine learning algorithms
  \item Heatmap visualizations
}
The matrix format preserves the relationship between amino acids and their 
synonymous codons while enabling sample-wise comparisons.
}
\examples{
\donttest{
# Using built-in data
data("prepared_fasta", package = "RSCUcaller")
rscu_data <- get_RSCU(prepared_fasta)
matrix_result <- get_matrix(rscu_data)

# Access matrix components
heatmap_matrix <- matrix_result$matrix
df_for_analysis <- matrix_result$df_matrix
}

}
\seealso{
Related functions:
\code{\link{get_RSCU}} for generating input data,
\code{\link{PCA_RSCU}} for dimensionality reduction,
\code{\link{heatmap_RSCU}} for visualization
}
