% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RSCU.R
\name{get_RSCU}
\alias{get_RSCU}
\title{Calculating RSCU for multiple sequences}
\usage{
get_RSCU(merged_sequences = "")
}
\arguments{
\item{merged_sequences}{path to the file with prepared sequences via the prepare_fasta() function}
}
\value{
A `data.frame` object containing Rscu_all.
}
\description{
Calculating RSCU for multiple sequences
}
