% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RSCU_other.R
\name{get_RSCU_other2}
\alias{get_RSCU_other2}
\title{Calculating Relative Synonymous Codon Usage (RSCU) based on table}
\usage{
get_RSCU_other2(
  merged_sequences = "",
  pseudo_count = 1,
  samples_table = samples_table
)
}
\arguments{
\item{merged_sequences}{A path to sequences or a list of sequences.}

\item{pseudo_count}{A value added to codon counts to avoid division by zero.}

\item{samples_table}{table with sequence descriptions and codon_table_id - an integer specifying the codon table (1-6,9-16,21-26).}
}
\value{
A data frame with columns: codon, amino acid, eff, RSCU.
}
\description{
This function calculates RSCU values for codons in a given nucleotide sequence.
}
