% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_RSCU.R
\name{heatmap_RSCU}
\alias{heatmap_RSCU}
\title{Creating a heatmap and grouping dendrogram based on the table with the get_RSCU function}
\usage{
heatmap_RSCU(heatmap_color = "", get_RSCU_out = get_RSCU_out, select = "")
}
\arguments{
\item{heatmap_color}{Selecting colors for the heatmap from among "red_green","green_red","blue_green","green_blue","blue_red","red_blue"}

\item{get_RSCU_out}{the table was created using the get_RSCU function}

\item{select}{the "heatmap" or "dendrogram" output}
}
\value{
A plot object.
}
\description{
Creating a heatmap and grouping dendrogram based on the table with the get_RSCU function
}
