% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_RSCU_double.R
\name{histogram_RSCU_double}
\alias{histogram_RSCU_double}
\title{Creating a double barplot based on the table with the get_RSCU function}
\usage{
histogram_RSCU_double(
  get_RSCU_out_left = get_RSCU_out,
  get_RSCU_out_right = get_RSCU_out,
  title_left = "",
  title_right = ""
)
}
\arguments{
\item{get_RSCU_out_left}{the table was created using the get_RSCU function, which will be drawn on the left side of the chart.}

\item{get_RSCU_out_right}{the table was created using the get_RSCU function, which will be drawn on the right side of the chart.}

\item{title_left}{left chart title}

\item{title_right}{right chart title}
}
\value{
A plot object.
}
\description{
Creating a double barplot based on the table with the get_RSCU function
}
