% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_RSCU.R
\name{PCA_RSCU}
\alias{PCA_RSCU}
\title{Creating a PCA based on the output from get_matrix function}
\usage{
PCA_RSCU(get_matrix_out = get_matrix_out, grouping_table = grouping_table)
}
\arguments{
\item{get_matrix_out}{the table was created using the get_RSCU function}

\item{grouping_table}{the table with Species and group column names}
}
\value{
A plot object.
}
\description{
Creating a PCA based on the output from get_matrix function
}
