% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepared_fasta.R
\docType{data}
\name{prepared_fasta}
\alias{prepared_fasta}
\title{FASTA sequences corresponding to samples_table after get_RSCU function}
\format{
A set of FASTA files.
}
\source{
Internal dataset for microbiome analysis.
}
\usage{
data(prepared_fasta)
}
\value{
A fasta list.
}
\description{
This dataset consists of FASTA files containing mitogenome gene sequences corresponding to the samples listed in the samples_table.
Each FASTA file is named according to the ID in the samples_table.
}
\examples{
data(prepared_fasta)
print(prepared_fasta)
}
\keyword{datasets}
