% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_scat_box.R
\name{stat_scat_box}
\alias{stat_scat_box}
\title{Creating a csv table, boxplots and scatter plots table based on the table with the get_RSCU function}
\usage{
stat_scat_box(get_RSCU_out, width, height, res, p.adjust.method = "bonferroni")
}
\arguments{
\item{get_RSCU_out}{the table was created using the get_RSCU function}

\item{width}{width of charts}

\item{height}{height of charts}

\item{res}{resolution of charts}

\item{p.adjust.method}{the selection of a p-value correction method: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr"}
}
\value{
A table with statistics, boxplots, barplots and scatter plots.
}
\description{
Creating a csv table, boxplots and scatter plots table based on the table with the get_RSCU function
}
